feat: wasm-rasterizer first draft (its way slower than js implementation :)

+++ src/rust/rasterizer/src/scripts.rs
@@ -1,10 +1,10 @@
 pub fn get_pixel_data(
     texture_width: i32,
     texture_height: i32,
-    texture_data: &Vec<i32>,
+    texture_data: &Vec<u8>,
     u: f64,
     v: f64
-) -> (i32, i32, i32, i32) {
+) -> (u8, u8, u8, u8) {
     let x: i32 = (v * texture_width as f64) as i32;
     let y: i32 = (u * texture_height as f64) as i32;
